test: move integration tests to dedicated folder
